+++ rindti/models/dti/baseline/__init__.py
@@ -1,0 +1,5 @@
+from .base_baseline import BaseBaseline
+from .max_likelihood import Max
+from .prot_drug_max_likelihood import ProtDrugMax
+
+__all__ = ["BaseBaseline", "Max", "ProtDrugMax"]